+++ src/test/java/org/bigml/binding/ComputeMultivotePredictionsStepdefs.java
@@ -11,9 +11,9 @@ import org.json.simple.JSONValue;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import cucumber.annotation.en.Given;
-import cucumber.annotation.en.Then;
-import cucumber.annotation.en.When;
+import io.cucumber.java.en.Given;
+import io.cucumber.java.en.Then;
+import io.cucumber.java.en.When;
 
 public class ComputeMultivotePredictionsStepdefs {