+++ src/test/java/org/bigml/binding/DeepnetsStepdefs.java
@@ -1,11 +1,12 @@
 package org.bigml.binding;
 
-import cucumber.annotation.en.Given;
 import org.json.simple.JSONObject;
 import org.json.simple.JSONValue;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
+
+import io.cucumber.java.en.Given;
 
 
 public class DeepnetsStepdefs {