+++ src/test/java/org/bigml/binding/ExecutionsStepdefs.java
@@ -1,7 +1,5 @@
 package org.bigml.binding;
 
-import cucumber.annotation.en.Given;
-import cucumber.annotation.en.Then;
 import org.json.simple.JSONArray;
 import org.json.simple.JSONObject;
 import org.slf4j.Logger;
@@ -13,6 +11,9 @@ import java.util.*;
 import org.bigml.binding.utils.Utils;
 
 import static org.junit.Assert.*;
+
+import io.cucumber.java.en.Given;
+import io.cucumber.java.en.Then;
 
 public class ExecutionsStepdefs {