+++ src/test/java/org/bigml/binding/LibrariesStepdefs.java
@@ -1,16 +1,14 @@
 package org.bigml.binding;
 
-import cucumber.annotation.en.Given;
-import cucumber.annotation.en.Then;
-
 import org.json.simple.JSONObject;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 
-import java.util.*;
-
 import static org.junit.Assert.*;
+
+import io.cucumber.java.en.Given;
+import io.cucumber.java.en.Then;
 
 public class LibrariesStepdefs {