+++ src/test/java/org/bigml/binding/OrganizationStepdefs.java
@@ -4,8 +4,8 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 
-import cucumber.annotation.After;
-import cucumber.annotation.Before;
+import io.cucumber.java.After;
+import io.cucumber.java.Before;
 
 public class OrganizationStepdefs {