+++ src/test/java/org/bigml/binding/TopicModelsStepdefs.java
@@ -1,9 +1,5 @@
 package org.bigml.binding;
 
-import cucumber.annotation.en.Given;
-import cucumber.annotation.en.Then;
-import cucumber.annotation.en.When;
-
 import org.json.simple.JSONArray;
 import org.json.simple.JSONObject;
 import org.json.simple.JSONValue;
@@ -18,8 +14,11 @@ import java.util.*;
 
 import static org.junit.Assert.*;
 
-import org.bigml.binding.LocalTopicModel;
 import org.bigml.binding.utils.Utils;
+
+import io.cucumber.java.en.Given;
+import io.cucumber.java.en.Then;
+import io.cucumber.java.en.When;
 
 public class TopicModelsStepdefs {